bot/requirements.txt
Flask==1.0.2
google-cloud-storage==1.13.0
google-cloud-logging==1.8.0
google-cloud-datastore==1.7.1
requests==2.19.1
python-telegram-bot==11.1.0
TwitterAPI==2.5.4
grpcio
pylint
python-dotenv